# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "3DManager API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/notifications/unread": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationObjectListCommonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationObjectListCommonResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationObjectListCommonResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanCommonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanCommonResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanCommonResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BooleanCommonResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorProperties"
          },
          "data": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorProperties": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NotificationObject": {
        "type": "object",
        "properties": {
          "notificationId": {
            "type": "integer",
            "format": "int32"
          },
          "notificationIsRead": {
            "type": "boolean"
          },
          "notificationMessage": {
            "type": "string",
            "nullable": true
          },
          "notificationRegisterDate": {
            "type": "string",
            "format": "date-time"
          },
          "notificationUserId": {
            "type": "integer",
            "format": "int32"
          },
          "notificationType": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "NotificationObjectListCommonResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorProperties"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationObject"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}